add private response for participants

+++ controllers/participants.controller.js
@@ -1,5 +1,6 @@
 import { validationResult } from 'express-validator';
 import Participants from '../models/participants.js';
+import checkAdmin from '../helpers/checkAdmin.js';
 
 class ParticipantsController {
     async createParticipant(req, res) {
@@ -21,12 +22,17 @@ class ParticipantsController {
 
     async readParticipant(req, res) {
         try {
-            const { id } = req.query;
-            const event = await Participants.findOne({ _id: id });
-            if (event) {
-                return res.json(event);
+            const { id, token } = req.query;
+            if (checkAdmin(token)) {
+                const event = await Participants.findOne({ _id: id });
+                if (event) {
+                    return res.json(event);
+                }
+                res.status(400).json({ message: 'Event not found' });
             }
-            res.status(400).json({ message: 'Event not found' });
+            res.status(400).json({
+                message: 'You have not permission for this response',
+            });
         } catch (err) {
             res.status(400).json({ message: err.message });
         }
@@ -34,12 +40,17 @@ class ParticipantsController {
 
     async eventsParticipant(req, res) {
         try {
-            const { eventId } = req.query;
-            const event = await Participants.find({ eventsId: eventId });
-            if (event) {
-                return res.json(event);
+            const { eventId, token } = req.query;
+            if (checkAdmin(token)) {
+                const event = await Participants.find({ eventsId: eventId });
+                if (event) {
+                    return res.json(event);
+                }
+                res.status(400).json({ message: 'Event not found' });
             }
-            res.status(400).json({ message: 'Event not found' });
+            res.status(400).json({
+                message: 'You have not permission for this response',
+            });
         } catch (err) {
             res.status(400).json({ message: err.message });
         }
@@ -47,27 +58,32 @@ class ParticipantsController {
 
     async updateParticipants(req, res) {
         try {
-            const { _id, eventID, data } = req.body;
-            const event = await Participants.findOne({ _id });
-            if (!event) {
-                return res.status(400).json({ message: 'Event not found' });
-            }
-
-            const filter = { _id };
-            const update = {
-                eventID,
-                data,
-            };
-
-            const newEvent = await Participants.findOneAndUpdate(
-                filter,
-                update,
-                {
-                    new: true,
+            const { _id, eventID, data, token } = req.body;
+            if (checkAdmin(token)) {
+                const event = await Participants.findOne({ _id });
+                if (!event) {
+                    return res.status(400).json({ message: 'Event not found' });
                 }
-            );
 
-            res.json(newEvent);
+                const filter = { _id };
+                const update = {
+                    eventID,
+                    data,
+                };
+
+                const newEvent = await Participants.findOneAndUpdate(
+                    filter,
+                    update,
+                    {
+                        new: true,
+                    }
+                );
+
+                res.json(newEvent);
+            }
+            res.status(400).json({
+                message: 'You have not permission for this response',
+            });
         } catch (err) {
             res.status(400).json({ message: err.message });
         }
@@ -75,15 +91,19 @@ class ParticipantsController {
 
     async deleteParticipant(req, res) {
         try {
-            const { _id } = req.body;
+            const { _id, token } = req.body;
+            if (checkAdmin(token)) {
+                const result = await Participants.deleteOne({ _id });
 
-            const result = await Participants.deleteOne({ _id });
+                if (result.deletedCount > 0) {
+                    return res.json({ message: 'Page was deleted' });
+                }
 
-            if (result.deletedCount > 0) {
-                return res.json({ message: 'Page was deleted' });
+                res.status(400).json({ message: 'Something going wrong' });
             }
-
-            res.status(400).json({ message: 'Something going wrong' });
+            res.status(400).json({
+                message: 'You have not permission for this response',
+            });
         } catch (err) {
             res.status(400).json({ message: err.message });
         }
